fix: use int16 for status code as per db schema

+++ pkg/types/tracedetailtypes/waterfall.go
@@ -63,7 +63,7 @@ type WaterfallSpan struct {
 	Resource           map[string]string `json:"resource"`
 	ResponseStatusCode string            `json:"response_status_code"`
 	SpanID             string            `json:"span_id"`
-	StatusCode         int32             `json:"status_code"`
+	StatusCode         int16             `json:"status_code"`
 	StatusCodeString   string            `json:"status_code_string"`
 	StatusMessage      string            `json:"status_message"`
 	TimeUnixMilli      uint64            `json:"timestamp"`
@@ -113,7 +113,7 @@ type SpanModel struct {
 	Flags              uint32             `ch:"flags"`
 	IsRemote           string             `ch:"is_remote"`
 	TraceState         string             `ch:"trace_state"`
-	StatusCode         int32              `ch:"status_code"`
+	StatusCode         int16              `ch:"status_code"`
 	DBName             string             `ch:"db_name"`
 	DBOperation        string             `ch:"db_operation"`
 	HTTPMethod         string             `ch:"http_method"`
